+++ LeafWeighing.Application/DTOs/Deduction/SaveDeductionRequestDto.cs
@@ -6,11 +6,11 @@ public class SaveDeductionRequestDto
     public string? SupplierName { get; set; }
     public string? Route { get; set; }
     public string? LeafType { get; set; }
-    public decimal BagWeight { get; set; }
-    public decimal Coarse { get; set; }
-    public decimal Water { get; set; }
-    public decimal Boiled { get; set; }
-    public decimal Rejected { get; set; }
+    public int BagWeight { get; set; }
+    public int Coarse { get; set; }
+    public int Water { get; set; }
+    public int Boiled { get; set; }
+    public int Rejected { get; set; }
     public string? UserName { get; set; }
     public string? Mode { get; set; }
     public string? PcName { get; set; }
